remove Mantine prefix
