+++ packages/frontend/src/const/routes.ts
@@ -1,12 +1,11 @@
 export const ROUTE_LANDING = "/landing";
-export const ROUTE_MANTINE_HOME = "/mantine";
 export const ROUTE_HOME = "/home";
-export const ROUTE_HOME_PLAY_QUEUE = "/mantine/play_queue";
-export const ROUTE_HOME_BROWSER = "/mantine/browser";
-export const ROUTE_HOME_PLAYLIST = "/mantine/playlists";
-export const ROUTE_HOME_FILE_EXPLORE = "/mantine/file_explore";
-export const ROUTE_HOME_SEARCH = "/mantine/search";
-export const ROUTE_HOME_PLUGIN = "/mantine/plugin";
-export const ROUTE_HOME_SETTINGS = "/mantine/settings";
-export const ROUTE_HOME_ALL_SONGS = "/mantine/all_songs";
-export const ROUTE_HOME_RECENTLY_ADDED = "/mantine/recently_added";
+export const ROUTE_HOME_PLAY_QUEUE = "/home/play_queue";
+export const ROUTE_HOME_BROWSER = "/home/browser";
+export const ROUTE_HOME_PLAYLIST = "/home/playlists";
+export const ROUTE_HOME_FILE_EXPLORE = "/home/file_explore";
+export const ROUTE_HOME_SEARCH = "/home/search";
+export const ROUTE_HOME_PLUGIN = "/home/plugin";
+export const ROUTE_HOME_SETTINGS = "/home/settings";
+export const ROUTE_HOME_ALL_SONGS = "/home/all_songs";
+export const ROUTE_HOME_RECENTLY_ADDED = "/home/recently_added";
