+++ packages/frontend/src/features/global_filter/components/GlobalFilterBox.tsx
@@ -1,15 +1,7 @@
-import {
-	Box,
-	CloseButton,
-	Icon,
-	Input,
-	InputGroup,
-	InputLeftElement,
-	InputRightElement,
-} from "@chakra-ui/react";
 import { useRef } from "react";
-import { IoSearchOutline } from "react-icons/io5";
 
+import { ActionIcon, Group, TextInput } from "@mantine/core";
+import { IconSearch, IconX } from "@tabler/icons-react";
 import { useHandleGlobalFilterTextChangeWithDebounce } from "../hooks/useHandleGlobalFilterTextChangeWithDebounce";
 
 /**
@@ -40,34 +32,36 @@ export function GlobalFilterBox() {
 
 	return (
 		<>
-			<Box w="35%" minW="100px" maxW="500px">
-				<InputGroup>
-					<InputLeftElement
-						pointerEvents="none"
-						color="gray.500"
-						fontSize="1.2em"
-					>
-						<Icon as={IoSearchOutline} />
-					</InputLeftElement>
-					<Input
-						ref={inputRef}
-						placeholder="Filter songs..."
-						onChange={(e) => {
-							handleGlobalFilterTextChange(e.target.value);
-						}}
-					/>
-					<InputRightElement color="gray.500" fontSize="1.2em">
-						<CloseButton
+			<Group w="35%" miw="100px" maw="500px">
+				<TextInput
+					ref={inputRef}
+					size="md"
+					placeholder="Filter songs..."
+					leftSection={
+						<ActionIcon variant="transparent" color="gray.5">
+							<IconSearch />
+						</ActionIcon>
+					}
+					leftSectionPointerEvents="none"
+					onChange={(e) => {
+						handleGlobalFilterTextChange(e.target.value);
+					}}
+					rightSection={
+						<ActionIcon
+							variant="transparent"
+							color="gray.5"
 							onClick={() => {
 								if (inputRef.current != null) {
 									inputRef.current.value = "";
 								}
 								handleGlobalFilterTextChange("");
 							}}
-						/>
-					</InputRightElement>
-				</InputGroup>
-			</Box>
+						>
+							<IconX />
+						</ActionIcon>
+					}
+				/>
+			</Group>
 		</>
 	);
 }
