--- packages/frontend/src/features/play_queue/components/MantinePlayQueue.tsx
@@ -1,82 +1,0 @@
-import type { SongTableColumn } from "@sola_mpd/domain/src/models/song_table_pb.js";
-import { useCallback, useState } from "react";
-
-import { Box } from "@mantine/core";
-import { UpdateMode } from "../../../types/stateTypes";
-import { CenterSpinner } from "../../loading";
-import { PlaylistSelectModal, usePlaylistSelectModal } from "../../playlist";
-import {
-	ColumnEditModal,
-	useColumnEditModalProps,
-	useSongTableState,
-	useUpdateSongTableState,
-} from "../../song_table";
-import { MantineSongTable } from "../../song_table/components/MantineSongTable";
-import { usePlayQueueSongTableProps } from "../hooks/usePlayQueueSongTableProps";
-
-/**
- * Displays and manages the MPD play queue.
- *
- * Renders a customizable song table with support for selection,
- * context menu actions, column editing, and playlist integration.
- * Includes loading states and modals for user interactions.
- *
- * Supports drag-and-drop reordering and multi-song operations
- * through an interactive table interface.
- *
- * @returns Play queue component with table and modals
- */
-export function MantinePlayQueue() {
-	const songTableState = useSongTableState();
-	const updateSongTableState = useUpdateSongTableState();
-
-	const [isColumnEditModalOpen, setIsColumnEditModalOpen] = useState(false);
-
-	const {
-		songsToAddToPlaylistRef,
-		setIsPlaylistSelectModalOpen,
-		playlistSelectModalProps,
-	} = usePlaylistSelectModal();
-
-	const songTableProps = usePlayQueueSongTableProps(
-		songsToAddToPlaylistRef,
-		setIsPlaylistSelectModalOpen,
-		setIsColumnEditModalOpen,
-	);
-
-	const onColumnsUpdated = useCallback(
-		async (columns: SongTableColumn[]) => {
-			if (songTableState === undefined) {
-				return;
-			}
-			const newSongTableState = songTableState.clone();
-			newSongTableState.columns = columns;
-			await updateSongTableState(
-				newSongTableState,
-				UpdateMode.LOCAL_STATE | UpdateMode.PERSIST,
-			);
-		},
-		[songTableState, updateSongTableState],
-	);
-	const columnEditModalProps = useColumnEditModalProps(
-		isColumnEditModalOpen,
-		songTableState?.columns || [],
-		setIsColumnEditModalOpen,
-		onColumnsUpdated,
-		async () => {},
-	);
-
-	if (songTableProps === undefined || columnEditModalProps === undefined) {
-		return <CenterSpinner />;
-	}
-
-	return (
-		<>
-			<Box w="100%" h="100%">
-				<MantineSongTable {...songTableProps} />
-				<PlaylistSelectModal {...playlistSelectModalProps} />
-				<ColumnEditModal {...columnEditModalProps} />
-			</Box>
-		</>
-	);
-}
