+++ packages/frontend/src/features/play_queue/components/PlayQueue.tsx
@@ -1,7 +1,7 @@
-import { Box } from "@chakra-ui/react";
 import type { SongTableColumn } from "@sola_mpd/domain/src/models/song_table_pb.js";
 import { useCallback, useState } from "react";
 
+import { Box } from "@mantine/core";
 import { UpdateMode } from "../../../types/stateTypes";
 import { CenterSpinner } from "../../loading";
 import { PlaylistSelectModal, usePlaylistSelectModal } from "../../playlist";
@@ -67,7 +67,7 @@ export function PlayQueue() {
 	);
 
 	if (songTableProps === undefined || columnEditModalProps === undefined) {
-		return <CenterSpinner className="layout-border-top layout-border-left" />;
+		return <CenterSpinner />;
 	}
 
 	return (
