--- packages/frontend/src/features/profile/components/MantineMpdProfileSelector.tsx
@@ -1,58 +1,0 @@
-import { Group, Select } from "@mantine/core";
-import { useNotificationMantine } from "../../../lib/chakra/hooks/useNotificationMantine";
-import { useEnabledOutputDevice } from "../../output_devices";
-import { useChangeCurrentMpdProfile } from "../hooks/useChangeCurrentMpdProfile";
-import { useMpdProfileState } from "../states/mpdProfileState";
-
-/**
- * Dropdown for MPD profile selection.
- *
- * Shows available profiles with current output device.
- *
- * @returns Selection component
- */
-export function MantineMpdProfileSelector() {
-	const notify = useNotificationMantine();
-
-	const mpdProfileState = useMpdProfileState();
-	const enabledOutputDevice = useEnabledOutputDevice();
-	const changeCurrentMpdProfile = useChangeCurrentMpdProfile();
-
-	return (
-		<>
-			<Group px={0} miw="100px" maw="300px">
-				{mpdProfileState?.currentProfile === undefined ||
-				enabledOutputDevice === undefined ? (
-					<Select size="md" placeholder="Loading profiles..." />
-				) : (
-					<Select
-						size="md"
-						value={mpdProfileState.currentProfile.name}
-						data={mpdProfileState.profiles.map((profile) => {
-							const isSelected =
-								profile.name === mpdProfileState.currentProfile?.name;
-							const text =
-								profile.name +
-								(isSelected ? ` - ${enabledOutputDevice?.name}` : "");
-							return {
-								value: profile.name,
-								label: text,
-							};
-						})}
-						onChange={async (value) => {
-							if (value == null) {
-								return;
-							}
-							await changeCurrentMpdProfile(value);
-							notify({
-								status: "info",
-								title: "MPD profile changed",
-								description: `MPD profile is changed to ${value}`,
-							});
-						}}
-					/>
-				)}
-			</Group>
-		</>
-	);
-}
