--- packages/frontend/src/pages/home/MantinePlayQueuePage.tsx
@@ -1,9 +1,0 @@
-import { MantinePlayQueue } from "../../features/play_queue/components/MantinePlayQueue";
-
-export function MantinePlayQueuePage() {
-	return (
-		<>
-			<MantinePlayQueue />
-		</>
-	);
-}
